fix: move projects data to separate file
